Refactor trip Dashboard to be named Timeline
